Steps toward generalizing renderable assets
