+++ runtime/functor-runtime-common/src/texture/texture_data.rs
@@ -1,4 +1,8 @@
-use super::PixelFormat;
+use glow::HasContext;
+
+use crate::asset::RenderableAsset;
+
+use super::{PixelFormat, Texture2D, TextureOptions};
 
 #[derive(Clone)]
 pub struct TextureData {
@@ -30,6 +34,57 @@ impl TextureData {
             width,
             height,
             format: PixelFormat::RGBA,
+        }
+    }
+}
+
+impl RenderableAsset for TextureData {
+    type HydratedType = glow::Texture;
+    type OptionsType = TextureOptions;
+
+    fn load(&self, gl_context: &glow::Context, options: &Self::OptionsType) -> Self::HydratedType {
+        unsafe {
+            let gl = gl_context;
+            let texture = gl.create_texture().expect("Texture to be created");
+            gl.bind_texture(glow::TEXTURE_2D, Some(texture));
+
+            // Set texture parameters
+            let wrap_val = if options.wrap {
+                glow::REPEAT
+            } else {
+                glow::CLAMP_TO_EDGE
+            };
+
+            let filter = if options.linear {
+                glow::LINEAR
+            } else {
+                glow::NEAREST
+            };
+
+            gl.tex_parameter_i32(glow::TEXTURE_2D, glow::TEXTURE_WRAP_S, wrap_val as i32);
+            gl.tex_parameter_i32(glow::TEXTURE_2D, glow::TEXTURE_WRAP_T, wrap_val as i32);
+            gl.tex_parameter_i32(glow::TEXTURE_2D, glow::TEXTURE_MIN_FILTER, filter as i32);
+            gl.tex_parameter_i32(glow::TEXTURE_2D, glow::TEXTURE_MAG_FILTER, filter as i32);
+
+            let format = match self.format {
+                PixelFormat::RGB => glow::RGB,
+                PixelFormat::RGBA => glow::RGBA,
+            };
+
+            gl.tex_image_2d(
+                glow::TEXTURE_2D,
+                0,
+                glow::RGBA as i32,
+                self.width as i32,
+                self.height as i32,
+                0,
+                format,
+                glow::UNSIGNED_BYTE,
+                Some(&self.bytes),
+            );
+
+            gl.bind_texture(glow::TEXTURE_2D, None);
+            texture
         }
     }
 }
